Add client reconnection helper
This moves in a direction seemingly opposite the recent move toward
simplifying the API in that there are now two ways to set client keys.
The best way to think of it is that `crypt_stream_set_keys_client` is
for low level code that wants to carefully manage memory. You will need
to set the keys before each and every connection in that case. For ease
of use, use the new connection method that takes the keys. That will
allow you to call `client_reconnect` without needing to lose sleep over
whether or not the keys are loaded.