Silently handle server key generation
On the first ever startup, it is unlikely that a fus.ini file will
exist. Due to all of the server to server connections fus uses,
encryption keys absolutely have to be available, so we will generate
those if the default (empty) value is detected. This prevents spurious
assertions from crashing us at worst and at best enabling a one command
shard.