Improve handling of decrypted clients
Clients can request decrypted connections by not sending the Y-data
buffer in the connection packet. This is great for debugging, but it is
a security risk for the server to server connections because the key
exchange effectively authenticates the connection. Therefore, we require
admin and db connections to be encrypted. Whether or not the player
facing services can be decrypted has been exposed as a compile-time
option.