Add stream peek support
Peeking works just like a read--you supply a complete net struct to
peek. When the next read is executed, the index into the net struct is
not reset and will resume there. This allows one to read in the message
header (type), then the message itself. The benefit is that we can now
use the same structures for both reading and writing messages.